better downloading with retry

+++ download_utils.py
@@ -4,24 +4,58 @@ import os
 import shutil
 import tqdm
 import requests
+import time
+from functools import wraps
+import traceback
+tqdm.monitor_interval = 0  # workaround for https://github.com/tqdm/tqdm/issues/481
 
 
+# https://www.saltycrane.com/blog/2009/11/trying-out-retry-decorator-python/
+def retry(ExceptionToCheck, tries=4, delay=3, backoff=2):
+    def deco_retry(f):
+
+        @wraps(f)
+        def f_retry(*args, **kwargs):
+            mtries, mdelay = tries, delay
+            while mtries > 1:
+                try:
+                    return f(*args, **kwargs)
+                except KeyboardInterrupt as e:
+                    raise e
+                except ExceptionToCheck as e:
+                    print("%s, retrying in %d seconds..." % (str(e), mdelay))
+                    traceback.print_exc()
+                    time.sleep(mdelay)
+                    mtries -= 1
+                    mdelay *= backoff
+            return f(*args, **kwargs)
+
+        return f_retry  # true decorator
+
+    return deco_retry
+
+
+@retry(Exception)
 def download_file(url, file_path):
     r = requests.get(url, stream=True)
     total_size = int(r.headers.get('content-length'))
+    bar = tqdm.tqdm_notebook(total=total_size, unit='B', unit_scale=True)
+    bar.set_description(os.path.split(file_path)[-1])
+    incomplete_download = False
     try:
-        with open(file_path, 'wb', buffering=16*1024*1024) as f:
-            bar = tqdm.tqdm_notebook(total=total_size, unit='B', unit_scale=True)
-            bar.set_description(os.path.split(file_path)[-1])
-            for chunk in r.iter_content(32 * 1024):
+        with open(file_path, 'wb', buffering=16 * 1024 * 1024) as f:
+            for chunk in r.iter_content(1 * 1024 * 1024):
                 f.write(chunk)
                 bar.update(len(chunk))
-    except Exception:
-        print("Download failed")
+    except Exception as e:
+        raise e
     finally:
-        if os.path.getsize(file_path) != total_size:
+        bar.close()
+        if os.path.exists(file_path) and os.path.getsize(file_path) != total_size:
+            incomplete_download = True
             os.remove(file_path)
-            print("Removed incomplete download")
+    if incomplete_download:
+        raise Exception("Incomplete download")
 
 
 def download_from_github(version, fn, target_dir):
